WebScrapper/requirements.txt
#Requirements
python3
request
beautifulsoup4
pandas  


